+++ filters/stripes.py
@@ -6,9 +6,9 @@ class Stripes:
     @classmethod
     def config(cls):
         return {
-            "Width": {"type": "numeric", "range": [1, 1024], "input": True, "default": 2},
-            "Intensity": {"type": "numeric", "range": [0, 255], "default": 10},
-            "Speed": {"type": "numeric", "range": [-200, 200], "default": 0}
+            "Width": {"type": "integer", "range": [1, 1024], "input": True, "default": 2},
+            "Intensity": {"type": "integer", "range": [0, 255], "default": 10},
+            "Speed": {"type": "integer", "range": [-200, 200], "default": 0}
         }
 
     def __init__(self, width=5, intensity=10.0, speed=1, *args, **kwargs):
